==> scripts/0_StartWith_3_cleaning_final.ipynb <==
import os
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('/content/drive/MyDrive/EM:IP/최종 데이터 파일/variables/0_start_with_2_original_cleaning.csv', sep=',')
df['timestamp'] = pd.to_datetime(df['timestamp'])
df['time'] = df['time'].astype(float)
df

##연속중복 행 제거

df['reps'] = 0

for i in tqdm(range(len(df)-1)):
        if (df['ID'][i] == df['ID'][i+1]) and (df['instance'][i] == df['instance'][i+1]):
            df['reps'][i+1] = 1

# 연속으로 반복되는 행동이 아닌 경우 (reps=0)인 경우만 남김
df = df[(df['reps'] != 1)]
df = df.reset_index(drop = True)
df

df.to_csv('0_StartWith_3_cleaning_final.csv', index=False)
