==> scripts/1_GenerateVariables_1_time_abs.ipynb <==
import os
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('/content/drive/MyDrive/EM:IP/정리/최종 데이터 파일/variables/0_StartWith_4_final.csv', sep=',')
df['timestamp'] = pd.to_datetime(df['timestamp'])
df['time'] = df['time'].astype(float)
df

df_time = df[['ID','cnt','time','number','instance']]

stu_id = df['ID'].unique().tolist()
time_a1 = []
time_b1 = []
time_b2 = []
time_b3 = []
time_b4 = []
time_b5 = []
time_b6 = []
time_b7 = []
time_b8 = []
time_b9 = []
time_b10 = []
time_b11 = []
time_b12 = []
time_b13 = []
time_b14 = []
time_b15 = []
time_b16 = []
time_b17 = []
time_b18 = []
time_b19 = []
time_b20 = []
time_b21 = []
time_b22 = []
time_b23 = []
time_b24 = []
time_b25 = []
cnt = []

a1 = [2, 5, 7, 11]

b1 = [2, 5, 6, 9, 10]
b2 = [2, 5, 6, 9, 8, 9, 10]
b3 = [2, 5, 6, 9, 8, 9, 8, 9, 10]
b4 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 10]
b5 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 10]

b6 = [2, 5, 6, 8, 9, 10]
b7 = [2, 5, 6, 8, 9, 8, 9, 10]
b8 = [2, 5, 6, 8, 9, 8, 9, 8, 9, 10]
b9 = [2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 10]

b10 = [2, 5, 6, 9, 11]
b11 = [2, 5, 6, 9, 8, 11]
b12 = [2, 5, 6, 9, 8, 9, 11]
b13 = [2, 5, 6, 9, 8, 9, 8, 11]
b14 = [2, 5, 6, 9, 8, 9, 8, 9, 11]
b15 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 11]
b16 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 11]
b17 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 11]
b18 = [2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 11]

b19 = [2, 5, 6, 8, 9, 11]
b20 = [2, 5, 6, 8, 9, 8, 11]
b21 = [2, 5, 6, 8, 9, 8, 9, 11]
b22 = [2, 5, 6, 8, 9, 8, 9, 8, 11]
b23 = [2, 5, 6, 8, 9, 8, 9, 8, 9, 11]
b24 = [2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 11]
b25 = [2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 11]

city_subway 2
concession 5
daily 7
cancel 11
individual 6
trip_n 8
trip_4 9
Buy 10


1. a: city→concession→daily→cancel (2, 5, 7, 11)

2. b:
city-concession-individual-trip4-buy (2, 5, 6, 9, 10)
city-concession-individual-trip4-tripn-trip4-buy (2, 5, 6, 9, 8, 9, 10)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-buy (2, 5, 6, 9, 8, 9, 8, 9, 10)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-trip4-buy (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 10)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-trip4-tripn-trip4-buy (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 10)

city-concession-individual-trip4-cancel (2, 5, 6, 9, 11)
city-concession-individual-trip4-tripn	-cancel (2, 5, 6, 9, 8, 11)
city-concession-individual-trip4-tripn-trip4-cancel (2, 5, 6, 9, 8, 9, 11)
city-concession-individual-trip4-tripn-trip4-tripn-cancel (2, 5, 6, 9, 8, 9, 8, 11)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-cancel (2, 5, 6, 9, 8, 9, 8, 9, 11)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-cancel (2, 5, 6, 9, 8, 9, 8, 9, 8, 11)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-trip4-cancel (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 11)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-trip4-tripn-cancel (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 11)
city-concession-individual-trip4-tripn-trip4-tripn-trip4-tripn-trip4-tripn-trip4-cancel
(2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 11)

city-concession-individual-tripn-trip4-buy (2, 5, 6, 8, 9, 10)
city-concession-individual-tripn-trip4-tripn-trip4-buy (2, 5, 6, 8, 9, 8, 9, 10)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-trip4-buy (2, 5, 6, 8, 9, 8, 9, 8, 9, 10)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-trip4-tripn-trip4-buy (2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 10)

city-concession-individual-tripn-trip4-cancel (2, 5, 6, 8, 9, 11)
city-concession-individual-tripn-trip4-tripn-cancel (2, 5, 6, 8, 9, 8, 11)
city-concession-individual-tripn-trip4-tripn-trip4-cancel (2, 5, 6, 8, 9, 8, 9, 11)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-cancel (2, 5, 6, 8, 9, 8, 9, 8, 11)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-trip4-cancel (2, 5, 6, 8, 9, 8, 9, 8, 9, 11)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-trip4-tripn-cancel (2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 11)
city-concession-individual-tripn-trip4-tripn-trip4-tripn-trip4-tripn-trip4-cancel
(2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 11)

1 a: (2, 5, 7, 11)

b:
2 (2, 5, 6, 9, 10)
3 (2, 5, 6, 9, 8, 9, 10)
4 (2, 5, 6, 9, 8, 9, 8, 9, 10)
5 (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 10)
6 (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 10)

7 (2, 5, 6, 8, 9, 10)
8 (2, 5, 6, 8, 9, 8, 9, 10)
9 (2, 5, 6, 8, 9, 8, 9, 8, 9, 10)
10 (2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 10)

11 (2, 5, 6, 9, 11)
12 (2, 5, 6, 9, 8, 11)
13 (2, 5, 6, 9, 8, 9, 11)
14 (2, 5, 6, 9, 8, 9, 8, 11)
15 (2, 5, 6, 9, 8, 9, 8, 9, 11)
16 (2, 5, 6, 9, 8, 9, 8, 9, 8, 11)
17 (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 11)
18 (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 11)
19 (2, 5, 6, 9, 8, 9, 8, 9, 8, 9, 8, 9, 11)

20 (2, 5, 6, 8, 9, 11)
21 (2, 5, 6, 8, 9, 8, 11)
22 (2, 5, 6, 8, 9, 8, 9, 11)
23 (2, 5, 6, 8, 9, 8, 9, 8, 11)
24 (2, 5, 6, 8, 9, 8, 9, 8, 9, 11)
25 (2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 11)
26 (2, 5, 6, 8, 9, 8, 9, 8, 9, 8, 9, 11)

for k in tqdm(stu_id):

    df_time_t = df_time[df_time['ID'] == k]


    cnt_t = df_time_t.iloc[0,1]
    cnt.append(cnt_t)


    df_time_t2 = df_time_t.pivot(index='ID', columns='number', values=['instance'])

    time_a_t = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+4].tolist() == a1:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+4]['time'])
            diff = end-start
            time_a_t += diff
    time_a1.append(time_a_t)


    time_b_t1 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+5].tolist() == b1:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+5]['time'])
            diff = end-start
            time_b_t1 += diff
    time_b1.append(time_b_t1)


    time_b_t2 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+7].tolist() == b2:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+7]['time'])
            diff = end-start
            time_b_t2 += diff
    time_b2.append(time_b_t2)


    time_b_t3 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+9].tolist() == b3:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+9]['time'])
            diff = end-start
            time_b_t3 += diff
    time_b3.append(time_b_t3)


    time_b_t4 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+11].tolist() == b4:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+11]['time'])
            diff = end-start
            time_b_t4 += diff
    time_b4.append(time_b_t4)


    time_b_t5 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+13].tolist() == b5:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+13]['time'])
            diff = end-start
            time_b_t5 += diff
    time_b5.append(time_b_t5)


    time_b_t6 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+6].tolist() == b6:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+6]['time'])
            diff = end-start
            time_b_t6 += diff
    time_b6.append(time_b_t6)


    time_b_t7 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+8].tolist() == b7:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+8]['time'])
            diff = end-start
            time_b_t7 += diff
    time_b7.append(time_b_t7)


    time_b_t8 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+10].tolist() == b8:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+10]['time'])
            diff = end-start
            time_b_t8 += diff
    time_b8.append(time_b_t8)


    time_b_t9 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+12].tolist() == b9:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+12]['time'])
            diff = end-start
            time_b_t9 += diff
    time_b9.append(time_b_t9)


    time_b_t10 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+5].tolist() == b10:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+5]['time'])
            diff = end-start
            time_b_t10 += diff
    time_b10.append(time_b_t10)


    time_b_t11 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+6].tolist() == b11:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+6]['time'])
            diff = end-start
            time_b_t11 += diff
    time_b11.append(time_b_t11)


    time_b_t12 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+7].tolist() == b12:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+7]['time'])
            diff = end-start
            time_b_t12 += diff
    time_b12.append(time_b_t12)


    time_b_t13 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+8].tolist() == b13:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+8]['time'])
            diff = end-start
            time_b_t13 += diff
    time_b13.append(time_b_t13)


    time_b_t14 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+9].tolist() == b14:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+9]['time'])
            diff = end-start
            time_b_t14 += diff
    time_b14.append(time_b_t14)


    time_b_t15 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+10].tolist() == b15:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+10]['time'])
            diff = end-start
            time_b_t15 += diff
    time_b15.append(time_b_t15)


    time_b_t16 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+11].tolist() == b16:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+11]['time'])
            diff = end-start
            time_b_t16 += diff
    time_b16.append(time_b_t16)


    time_b_t17 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+12].tolist() == b17:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+12]['time'])
            diff = end-start
            time_b_t17 += diff
    time_b17.append(time_b_t17)


    time_b_t18 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+13].tolist() == b18:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+13]['time'])
            diff = end-start
            time_b_t18 += diff
    time_b18.append(time_b_t18)


    time_b_t19 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+6].tolist() == b19:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+6]['time'])
            diff = end-start
            time_b_t19 += diff
    time_b19.append(time_b_t19)


    time_b_t20 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+7].tolist() == b20:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+7]['time'])
            diff = end-start
            time_b_t20 += diff
    time_b20.append(time_b_t20)


    time_b_t21 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+8].tolist() == b21:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+8]['time'])
            diff = end-start
            time_b_t21 += diff
    time_b21.append(time_b_t21)


    time_b_t22 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+9].tolist() == b22:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+9]['time'])
            diff = end-start
            time_b_t22 += diff
    time_b22.append(time_b_t22)


    time_b_t23 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+10].tolist() == b23:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+10]['time'])
            diff = end-start
            time_b_t23 += diff
    time_b23.append(time_b_t23)


    time_b_t24 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+11].tolist() == b24:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+11]['time'])
            diff = end-start
            time_b_t24 += diff
    time_b24.append(time_b_t24)


    time_b_t25 = 0
    for i in range(max(df_time_t['number'])):
        if df_time_t2.iloc[0,i:i+12].tolist() == b25:
            start = float(df_time_t[df_time_t['number']==i+1]['time'])
            end = float(df_time_t[df_time_t['number']==i+12]['time'])
            diff = end-start
            time_b_t25 += diff
    time_b25.append(time_b_t25)

time_B1 = []
time_B2 = []
time_B3 = []
time_B4 = []

for i in range(len(stu_id)):
    B1 = time_b1[i]
    B2 = time_b2[i] + time_b3[i] + time_b4[i] + time_b5[i] + time_b6[i] + time_b7[i] + time_b8[i] + time_b9[i]
    B3 = time_b10[i]
    B4 = time_b11[i] + time_b12[i] + time_b13[i] + time_b14[i] + time_b15[i] + time_b16[i] + time_b17[i] + time_b18[i] + time_b19[i] + time_b20[i] + time_b21[i] + time_b22[i] + time_b23[i] + time_b24[i] + time_b25[i]
    time_B1.append(B1)
    time_B2.append(B2)
    time_B3.append(B3)
    time_B4.append(B4)

print(time_B1, time_B2,time_B3, time_B4)

time_a1

# start time
time_start = []

for i in tqdm(stu_id):
    df_t = df[df['ID'] == i]
    if len(df_t) >= 3:
        start = df_t.iloc[1,6] - df_t.iloc[0,6]
    else:
        start = 0
    time_start.append(start)

# end time
time_end = []

for i in tqdm(stu_id):
    df_t = df[df['ID'] == i]
    if len(df_t) >= 3:
        end = df_t.iloc[-1,6] - df_t.iloc[-2,6]
    else:
        end = 0
    time_end.append(end)

# total time
time_total = []

for i in tqdm(stu_id):
    df_t = df[df['ID'] == i]
    if len(df_t) >= 2:
        total = df_t.iloc[-2,6] - df_t.iloc[0,6]
    else:
        total = 0
    time_total.append(total)

# solving time
time_solving = []

for i in tqdm(stu_id):
    df_t = df[df['ID'] == i]
    if len(df_t) >= 4:
        solving = df_t.iloc[-2,6] - df_t.iloc[1,6]
    else:
        solving = 0
    time_solving.append(solving)

df_timeV_1_abs = pd.DataFrame({'ID':stu_id, 'cnt':cnt, 'time_a':time_a1,'time_b1':time_B1,'time_b2':time_B2,'time_b3':time_B3,'time_b4':time_B4,
                              'time_start':time_start, 'time_solving':time_solving, 'time_total':time_total})
#df.df_timeV_1_abs('ID', inplace=True)
df_timeV_1_abs

df_timeV_1_abs['time_answer'] = df_timeV_1_abs['time_a'] + df_timeV_1_abs['time_b1'] + df_timeV_1_abs['time_b2'] + df_timeV_1_abs['time_b3'] + df_timeV_1_abs['time_b4']
df_timeV_1_abs['time_irrelevent'] = df_timeV_1_abs['time_solving'] - df_timeV_1_abs['time_answer']
df_timeV_1_abs['time_explore_relevent'] = df_timeV_1_abs['time_a'] + df_timeV_1_abs['time_b3'] + df_timeV_1_abs['time_b4']
df_timeV_1_abs['time_explore_total'] = df_timeV_1_abs['time_solving'] - df_timeV_1_abs['time_b1'] + df_timeV_1_abs['time_b2']

# time_ticket_explore: b2와 b4에서 trip_n과 trip_4가 반복되는 시간

df_timeV_1_abs

df_timeV_1_abs.to_csv('1_GenerateVariables_1_time_abs.csv', index = False)
