==> scripts/1_GenerateVariables_1_time_mean.ipynb <==
import os
import numpy as np
import pandas as pd
from tqdm import tqdm
import warnings
warnings.filterwarnings('ignore')

from google.colab import drive
drive.mount('/content/drive')

df_original = pd.read_csv('/content/drive/MyDrive/EM:IP/정리/최종 데이터 파일/variables/0_StartWith_4_final.csv', sep=',')
df_time = pd.read_csv('/content/drive/MyDrive/EM:IP/정리/최종 데이터 파일/new_variable/1_GenerateVariables_1_time_abs.csv', sep=',')

stu_id = df_original['ID'].unique().tolist()

events_num = []

for i in tqdm(stu_id):
    df_t = df_original[df_original['ID'] == i]
    if max(df_t['number']) >= 3:
        length = max(df_t['number']) - 3
    else:
        length = 0
    events_num.append(length)

events_num

df_time['events_num'] = events_num

df_time

#df_time['time_avg_btw_explore'] = df_time['time_answer'] / df_time['length']
df_time['time_avg_btw_events'] = df_time['time_solving'] / df_time['events_num']

df_time

df_time.info()

df_time.to_csv('1_GenerateVariables_1_time_mean.csv', index = False)
