==> scripts/1_GenerateVariables_3_merged_outliers.ipynb <==
import os
import numpy as np
import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt

pip install scikit-learn-extra

from sklearn import preprocessing
from sklearn.preprocessing import RobustScaler
from sklearn_extra.cluster import KMedoids
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
import matplotlib.pyplot as plt
plt.rcParams['figure.figsize'] = (13, 8)
import seaborn as sns

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('/content/drive/MyDrive/EM:IP/최종 데이터 파일/variables/1_GenerateVariables_3_merge.csv', index_col = False)
df

df['cnt'] = df['cnt'].astype(str)
df['oecd'] = df['oecd'].astype(int)
df['credit'] = df['credit'].astype(int)
df['length'] = df['length'].astype(int)
df['time_start'] = df['time_start'].astype(float)
df['time_end'] = df['time_end'].astype(float)
df['time_solving'] = df['time_solving'].astype(float)
df['time_total'] = df['time_total'].astype(float)
df['time_answer'] = df['time_answer'].astype(float)
df['avg_time_btw_events'] = df['avg_time_btw_events'].astype(float)
df

df_1 = df[['ID', 'cnt', 'time_start', 'time_end', 'time_answer', 'time_not_answer','time_solving', 'time_total', 'events_num', 'avg_time_btw_events', 'length', 'credit', 'oecd']]
df_1.set_index('ID', inplace=True)
df_1

df_1.to_csv('2_ExploratoryAnalysis_0_start_with_all.csv', index=True)

df_2 = df_1[df_1['credit'] == 1]

df_2.to_csv('2_ExploratoryAnalysis_0_start_with_full_credit.csv', index=True)
